Moved additional request code into Faraday middlware
