+++ lib/moo_moo.rb
@@ -1,18 +1,19 @@
 require 'moo_moo/exceptions'
 
 module MooMoo
-  autoload :Version,        'moo_moo/version'
-  autoload :Config,         'moo_moo/config'
-  autoload :Command,        'moo_moo/command'
-  autoload :Response,       'moo_moo/response'
-  autoload :Base,           'moo_moo/base'
-  autoload :Lookup,         'moo_moo/lookup'
-  autoload :Nameserver,     'moo_moo/nameserver'
-  autoload :Provisioning,   'moo_moo/provisioning'
-  autoload :Transfer,       'moo_moo/transfer'
-  autoload :Cookie,         'moo_moo/cookie'
-  autoload :OpenSRSErrors,  'moo_moo/middleware/open_srs_errors'
-  autoload :ParseOpenSRS,  'moo_moo/middleware/parse_open_srs'
+  autoload :Version,           'moo_moo/version'
+  autoload :Config,            'moo_moo/config'
+  autoload :Command,           'moo_moo/command'
+  autoload :Response,          'moo_moo/response'
+  autoload :Base,              'moo_moo/base'
+  autoload :Lookup,            'moo_moo/lookup'
+  autoload :Nameserver,        'moo_moo/nameserver'
+  autoload :Provisioning,      'moo_moo/provisioning'
+  autoload :Transfer,          'moo_moo/transfer'
+  autoload :Cookie,            'moo_moo/cookie'
+  autoload :OpenSRSErrors,     'moo_moo/middleware/open_srs_errors'
+  autoload :ParseOpenSRS,      'moo_moo/middleware/parse_open_srs'
+  autoload :OpenSRSXMLBuilder, 'moo_moo/middleware/open_srs_xml_builder'
 
   class << self
     attr_accessor :config
